# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Secure Online Election Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /election-requests:
    get:
      operationId: listElectionRequests
      tags: [election-requests]
      summary: List election creator requests (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Election requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionRequestListResponse"
    post:
      operationId: createElectionRequest
      tags: [election-requests]
      summary: Submit election creator request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectionRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionRequest"
  /election-requests/{id}:
    get:
      operationId: getElectionRequest
      tags: [election-requests]
      summary: Get election request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Election request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionRequest"
  /election-requests/{id}/approve:
    post:
      operationId: approveElectionRequest
      tags: [election-requests]
      summary: Approve election creator request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionRequest"
  /election-requests/{id}/reject:
    post:
      operationId: rejectElectionRequest
      tags: [election-requests]
      summary: Reject election creator request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequestInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionRequest"

  # ─── ELECTIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, email, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, election_creator, voter]
        status:
          type: string
          enum: [active, blocked, pending]
        orgName:
          type: ["string", "null"]
        electionPurpose:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ElectionRequest:
      type: object
      required: [id, userId, orgName, electionPurpose, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        orgName:
          type: string
        electionPurpose:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ElectionRequestInput:
      type: object
      required: [orgName, electionPurpose]
      properties:
        orgName:
          type: string
        electionPurpose:
          type: string
    RejectRequestInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    ElectionRequestListResponse:
      type: object
      required: [requests, total, page, limit]
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/ElectionRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── ELECTION ────────────────────────────────────────────────────────────
...